Use only Pull Requests for patch creation.
Using only patches is enough, however to dodge any patch clashes the git
rebase was used to do the heavy lifting.

Moreover, when doing a git rebase for PR#415 when a merge conflict
arises the lib/protobuf/message.rb needs to have both changes manually
merged into one.
Also apparently one line in test suite rises up with a
merge conflict as well and it needs to have `.force_encoding` on the
conflicting line.

Note: there is a gotcha with another test where git does not make a
proper diff since it thinks its a binary file. Git is wrong about this
and an extra patch was added to compensate for this.